requirements.txt
discord.py>=2.3.2
anthropic>=0.7.2
python-dotenv>=1.0.0